fix(lens): hoist one fenced bridgeStream into @kolu/terminal-dag
Move the tap-pump primitive (carrying local.ts's per-event try/catch fence) into @kolu/terminal-dag and have both LocalTerminalEndpoint and kolu-watcher import it; delete the un-fenced copy in kolu-watcher/streamBridge.ts. The watcher's taps can no longer be silently frozen by one throwing event, and the two copies can't drift.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-6, raised by hickey). Not pushed or merged.
