+++ packages/kolu-watcher/src/streamBridge.ts
@@ -1,37 +1,16 @@
 /**
- * Stream adapters the watcher server needs to bridge between three stream
- * shapes: kaval's client tap streams, the DAG's in-memory channels, and the
- * watcher's own served surface streams.
+ * Stream adapters the watcher server needs to bridge between two of its stream
+ * shapes: the watcher's own served surface streams (`forwardStream`) and the
+ * change-tick streams it re-serves from kolu-git (`tickStream`). The tap-pump
+ * primitive that feeds kaval's taps onto the DAG's provider channels
+ * (`bridgeStream`, carrying the per-event fence) is shared with kolu-server via
+ * `@kolu/terminal-dag` — both hosts run the same provider DAG, so one copy.
  *
- * All three honour the same teardown contract: an aborted `signal` is EXPECTED
+ * Both honour the same teardown contract: an aborted `signal` is EXPECTED
  * (a kill / link drop), never an error — the kaval client streams are opened
  * with `{ signal }`, so an abort ends them cleanly and surfaces here as a
  * swallowed throw, not a failure.
  */
-
-/** Pump a (possibly promised) async-iterable into a callback until it ends or
- *  `signal` aborts. Fire-and-forget: used to feed kaval's taps onto the DAG's
- *  provider channels. `onError` fires only for a NON-abort failure (an abort is
- *  expected teardown). Mirrors kolu-server's `LocalTerminalEndpoint.bridgeStream`. */
-export function bridgeStream<T>(
-  source: AsyncIterable<T> | PromiseLike<AsyncIterable<T>>,
-  signal: AbortSignal,
-  onEvent: (value: T) => void,
-  onError?: (err: unknown) => void,
-): void {
-  void (async () => {
-    try {
-      const iterable = await source;
-      for await (const value of iterable) {
-        if (signal.aborted) return;
-        onEvent(value);
-      }
-    } catch (err) {
-      if (signal.aborted) return;
-      onError?.(err);
-    }
-  })();
-}
 
 /** Re-yield a (possibly promised) async-iterable as the watcher's own served
  *  stream, stopping on `signal` abort. Used to FORWARD kaval's pty taps
